LPS-57919 move constants to the API module
