+++ modules/addons/layout-type-controller/layout-type-controller-control-panel/src/com/liferay/layout/type/controller/control/panel/controller/ControlPanelLayoutTypeController.java
@@ -16,7 +16,7 @@ package com.liferay.layout.type.controller.control.panel.controller;
 
 import com.liferay.application.list.PanelAppRegistry;
 import com.liferay.application.list.PanelCategoryRegistry;
-import com.liferay.application.list.taglib.constants.ApplicationListWebKeys;
+import com.liferay.application.list.constants.ApplicationListWebKeys;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
 import com.liferay.portal.model.LayoutConstants;
 import com.liferay.portal.model.LayoutTypeController;
